119. 버그픽스: 다른사람 프로필에서 리트윗, 트윗 인용하기, 댓글 작성, 팔로잉에 있는 목록에서 팔로우나 언팔로우 할 경우 내 프로필처럼 작동함

+++ components/Main/Users/ProfileHeader.jsx
@@ -111,10 +111,14 @@ function ProfileHeader({ handleItemClick, activeItem }) {
     } catch (error) {
       console.error(error);
     }
-  }, [specificUser]);
+  }, [specificUser, currentUserId]);
 
   /*  언팔로우 */
   const handleUnfollowUser = useCallback(async () => {
+    if (!currentUserId) {
+      return alert("로그인이 필요합니다.");
+    }
+
     try {
       await userFunctions.unfollowUser(specificUser.id);
       dispatch(userActions.removeFollowing(specificUser.id));
@@ -122,7 +126,7 @@ function ProfileHeader({ handleItemClick, activeItem }) {
     } catch (error) {
       console.error(error);
     }
-  }, [specificUser]);
+  }, [specificUser, currentUserId]);
 
   return (
     <>
